Refactor SettingsModal into smaller feature-sliced components
